GraphView: Use new graphview and render background color.
