+++ src/widgets/GraphView.h
@@ -98,7 +98,7 @@ public:
 
 protected:
     std::unordered_map<ut64, GraphBlock> blocks;
-
+    QColor backgroundColor = QColor(Qt::white);
     // The vertical margin between blocks
     int block_vertical_margin = 32;
